+++ nonograma_core/Elementos_graficos/AssetManager.py
