+++ nonograma_core/Elementos_graficos/colores.py
