+++ nonograma_core/Logica/nonograma_numeros.py
