+++ nonograma_core/Logica/queueYStack.py
